Changement langage du module en TS
