+++ package.json
@@ -6,10 +6,24 @@
   "repository": "https://github.com/Natchi59/discordjs-handler.git",
   "author": "Natchi <nathan.caron@natchi.fr>",
   "license": "MIT",
+  "scripts": {
+    "devts": "tsnd --respawn --transpile-only ./test/index.ts",
+    "devjs": "nodemon ./test/index.js"
+  },
+  "keywords": [
+    "discordjs",
+    "discord.js",
+    "bot",
+    "bot handler",
+    "discordjs handler"
+  ],
   "dependencies": {
     "discord.js": "^12.5.3"
   },
-  "keywords": [
-    "discordjs", "discord.js", "bot", "bot handler", "discordjs handler"
-  ]
+  "devDependencies": {
+    "@types/node": "^16.3.3",
+    "nodemon": "^2.0.12",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "^4.3.5"
+  }
 }
